add index to consolidate common components
